@@ -10,9 +10,12 @@
   "author": "Zhongde Liu <zhongdeliu@googlemail.com>",
   "license": "MIT",
   "devDependencies": {
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
+    "babel-preset-env": "^1.6.0",
     "faker": "^4.1.0",
     "jquery": "^3.2.1",
-    "webpack": "^2.6.1",
+    "webpack": "^2.7.0",
     "webpack-dev-server": "^2.4.5"
   }
 }
